fix review comments
GIT_ORIGIN_SPP_REV_ID: 9d5abe566fc23944814cf86b3079da3603bf0247

+++ score/mw/log/slot_handle.h
@@ -26,7 +26,7 @@ namespace mw
 namespace log
 {
 
-using SlotIndex = std::uint16_t;
+using SlotIndex = std::uint32_t;
 
 /// \brief Represents an identifier that is owned by a `LogStream` to identify the correct memory slot where the
 /// streamed message shall be written.
